Implemented:
makeFieldsNoPrefix, classUnderscoreNoPrefixFields and classUnderscoreNoPrefixNamer

in order to support generation of overloaded field accessors (like makeFields) but without
the requirement to prefix each field name with the type name. That is, a field name
"_foo" will cause the overloaded field accessor "foo" to be created.

This might be the desired behaviour when the DuplicateRecordFields extension is used.