Fix -Wsimplifiable-class-constraints warning
On GHC HEAD, building `lens` warns with:

    src/Control/Lens/Internal/Exception.hs:193:10:warning: [-Wsimplifiable-class-constraints]
        • The constraint ‘Typeable (Handling a s m)’ matches
            a built-in instance
          This makes type inference for inner bindings fragile;
            either use MonoLocalBinds, or simplify it using the instance
        • In the instance declaration for ‘Exception (Handling a s m)’
        |
    193 | instance(Reifies s (SomeException -> Maybe a), Typeable (Handling a s m)) => Exception (Handling a s m) where
        |          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

This fixes the warning by simplfying the `Typeable (Handling a s m)`
content to just `(Typeable a, Typeable s, Typeable m)`. In the
process, I made the instance less poly-kinded than before, but this
is OK, since (1) the underlying `Handling` data type is not exported,
and (2) its only purpose is to be used in combination with
`reifyTypeable`, which doesn't make use of the previously poly-kinded
definition anyway.